+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix a zip deploy issue for the release workflow template (thanks to [domikuss](https://github.com/domikuss) for implementing this, see #220).
 - Fixed auto closing `*/` insertion for non docstring comments.
 - Fixed an issue where items would show a definition in string and comments (see #214).
+- Fixed an issue where methods definition would not point to the correct method.
 
 ## [3.1.0]